backend/requirements.txt
asgiref==3.8.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.1
sqlparse==0.5.3
tzdata==2025.2


